Fix train and test size calculation in create_split.py

+++ recipes/dcase2023_task4_baseline/utils/create_split.py
@@ -17,8 +17,8 @@ def split_df(df: pd.DataFrame, dur_df:pd.DataFrame, split_ratio: list, random_st
     csv = df.copy()
     csv.filename = csv.filename.apply(lambda x: x.split("/")[-1])
     print(f"Total: {len(csv)} files")
-    train_size = split_ratio[0] * len(csv)
-    test_size = split_ratio[1] * len(csv)
+    train_size = int(split_ratio[0] * len(csv))
+    test_size = int(split_ratio[1] * len(csv))
     
     train, temp = train_test_split(csv, train_size=train_size, random_state=random_state, stratify=df['event_label'])
     val, test = train_test_split(temp, test_size=test_size, random_state=random_state, stratify=temp['event_label'])
